Completing packaging and renaming components
