+++ src/ETMeta/core.py
